One generator, no chat agent, no legacy shortcode aliases in the manifest
Merges the entity, Start-operation and HTTP/OpenAPI generators into
`scripts/generator` behind a single `pnpm generate` / `generate:check`, with
one header, one seal and one staleness check. Deletes the in-app chat agent
(routes, hooks, contracts, workflows, MCP bridge) and the unrun `ai-eval`
harness. Strips `identifiers.legacy` from the manifest and generated tables;
`P-`/`L-` survive only as an inbound alias in `packages/shared`.

Lane C prerequisites: named `<Entity>ListItem` components, `native:` flags on
contracts and declarations replacing `native-operations.json`, generated
`APITypes.swift` aliases with `accessModifier: public`, `FieldDescriptor`
format/mobile metadata and a generated shortcode-prefix lookup, a `scan`
input arm resolved server-side, and integer metadata for whole-number
literals in the OpenAPI document.

+++ packages/schemas/src/problems.ts
@@ -1118,9 +1118,9 @@ const byTypeFields = Object.fromEntries(
 
 export const problemsCountSchema = z.object({
   /** Defect rows only — what the navbar badge and homepage banner show. */
-  total: z.number(),
+  total: z.number().int(),
   /** Coverage-backlog rows, reported separately so they never inflate `total`. */
-  coverageTotal: z.number(),
+  coverageTotal: z.number().int(),
   byType: z.object(byTypeFields),
 });
